requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
sqlalchemy==2.0.41
pydantic==2.10.6
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.2.1
httpx==0.28.1
python-multipart==0.0.9
